+++ frontend/src/components/account/favourites/Favourites.js
@@ -13,22 +13,22 @@
 // limitations under the License.
 
 import React, { Component } from "react";
-import AccountHeader from "./AccountHeader";
+import AccountHeader from "../account-header/AccountHeader";
 import Button from "react-bootstrap/Button";
 import { Link } from "react-router-dom";
-import { Recipe } from "../Recipe";
+import { Recipe } from "../../recipe/Recipe";
 import "./Favourites.css";
 
-class History extends Component {
+class Favourites extends Component {
   constructor(properties) {
     super(properties);
     this.state = {
-      recipes: [],
+      recipes: []
     };
   }
 
   componentDidMount() {
-    fetch("/api/history")
+    fetch("/api/favourites")
       .then((response) => response.json())
       .then((json) => this.setState({ recipes: json }))
       .catch((err) => console.log(err));
@@ -37,8 +37,8 @@ class History extends Component {
   render() {
     return (
       <div>
-        <AccountHeader />
-        <h1>History</h1>
+        <AccountHeader /> 
+        <h1>Favourites</h1>
         {this.state.recipes.map((recipe, index) => {
           const button = (
             <div className="right-side-btn">
@@ -47,6 +47,13 @@ class History extends Component {
                   Let's Go!
                 </Button>
               </Link>
+              <Button
+                className="remove-btn"
+                variant="danger"
+                onClick={this.removeFavourite}
+              >
+                Remove
+              </Button>
             </div>
           );
           return <Recipe key={index} recipe={recipe} buttons={button} />;
@@ -54,5 +61,19 @@ class History extends Component {
       </div>
     );
   }
+
+  removeFavourite = (recipe) => {
+    const recipeId = recipe.recipeId;
+    const request = new Request("/api/remove-favourite", {
+      method: "POST",
+      headers: {
+        "Content-Type": "application/json",
+        Accept: "application/json",
+      },
+      body: JSON.stringify(recipeId),
+    });
+    fetch(request).catch((err) => console.log(err));
+    window.location.reload();
+  }
 }
-export default History;
+export default Favourites;
