+++ frontend/src/components/home/Home.js
@@ -15,7 +15,7 @@
 import Button from "react-bootstrap/Button";
 import React, { Component } from "react";
 import { Link } from "react-router-dom";
-import account from "../icons/account.svg";
+import account from "../../icons/account.svg";
 import "./Home.css";
 
 class Home extends Component {
